% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTracksTools.R
\name{addBEADSmapq0TrackZcs}
\alias{addBEADSmapq0TrackZcs}
\title{addBEADSmapq0Track and z-score it}
\usage{
addBEADSmapq0TrackZcs(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
addBEADSmapq0Track and z-score it
}
\examples{
#parallel::mclapply(sprintf('REP\%.3i', 36:42)[-6], addBEADSmapq0TrackZcs), mc.cores = 8)
# sapply(sprintf('REP\%.3i', 3:24), addBEADSmapq0TrackZcs)
}
\seealso{
Other Peaks: \code{\link{addMapq0TrackZcs}},
  \code{\link{addMapq0Track}},
  \code{\link{addStrandedRNAseq}},
  \code{\link{callPeaksMACS}},
  \code{\link{combinePeaksIDR}},
  \code{\link{combinePeaksToBed}},
  \code{\link{combinePeaks}},
  \code{\link{enrichedRegionsCall}},
  \code{\link{jadb_addScaledTrack}},
  \code{\link{jadb_addTracksFromBAM}},
  \code{\link{meme_chip_local}}
}
\author{
Przemyslaw Stempor
}
