% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSops.R
\name{addExtractIDs}
\alias{addExtractIDs}
\title{Add extract to database using local CSV or google spreadsheet}
\usage{
addExtractIDs(csv, gsheet = TRUE)
}
\arguments{
\item{csv}{path to CSV or url to gsheet}

\item{gsheet}{is the source gsheet or local}
}
\description{
Accepts regular expressions.
}
\seealso{
Other FSops: \code{\link{addFilesFromBaseSpace}},
  \code{\link{addFilesFromCsv}},
  \code{\link{validateFilesFromBaseSpace}}
}
\author{
Przemyslaw Stempor
}
