% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSops.R
\name{addFilesFromBaseSpace}
\alias{addFilesFromBaseSpace}
\title{Add files to database using local CSV or google spreadsheet}
\usage{
addFilesFromBaseSpace(csv, root = "/mnt/jadb/DBfile/DBfiles",
  EXTABLE = "mydb.labexperiment", gsheet = TRUE, subset = NULL)
}
\arguments{
\item{csv}{path to CSV or url to gsheet}

\item{root}{root of the database FS}

\item{EXTABLE}{table to add the experiment to ("mydb.labrnaseq" for RNAseq)}

\item{gsheet}{is the source gsheet or local}
}
\description{
Accepts regular expressions.
}
\author{
Przemyslaw Stempor
}
\seealso{
Other FSops: \code{\link{addFilesFromCsv}},
  \code{\link{validateFilesFromBaseSpace}}
}

