% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSops.R
\name{addFilesFromCsv}
\alias{addFilesFromCsv}
\title{Add files to database using local CSV or google spreadsheet}
\usage{
addFilesFromCsv(csv, root = MOUNT, EXTABLE = "mydb.labexperiment",
  gsheet = TRUE)
}
\arguments{
\item{csv}{path to CSV or url to gsheet}

\item{root}{root of the database FS}

\item{EXTABLE}{table to add the experiment to ("mydb.labrnaseq" for RNAseq)}

\item{gsheet}{is the source gsheet or local}
}
\description{
Accepts regular expressions.
}
\seealso{
Other FSops: \code{\link{addExtractIDs}},
  \code{\link{addFilesFromBaseSpace}},
  \code{\link{validateFilesFromBaseSpace}}
}
\author{
Przemyslaw Stempor
}
