% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWtools.R
\name{addIDRtoJADB}
\alias{addIDRtoJADB}
\title{Adds replicates to database}
\usage{
addIDRtoJADB(IDs, res = 100L)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}

\item{res}{Resolution of tracks}

\item{outdir}{Directory to output BW file}
}
\description{
Adds replicates to database
}
\examples{
#combineReps(IDs)
}
\author{
Przemyslaw Stempor
}
\seealso{
Other BW: \code{\link{addRepToJADB}},
  \code{\link{combineReps}}
}

