% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTracksTools.R
\name{addMapq0Beads}
\alias{addMapq0Beads}
\title{addMapq0Beads}
\usage{
addMapq0Beads(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
addMapq0Beads
}
\examples{
#addMapq0Beads('AA001')
}
\seealso{
Other tracks: \code{\link{addMapq0BeadsRep}},
  \code{\link{addNonUniqueQ10Beads}},
  \code{\link{jadb_addAlignedBAM}}
}
\author{
Przemyslaw Stempor
}
