% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTracksTools.R
\name{addMapq0BeadsRep}
\alias{addMapq0BeadsRep}
\title{addMapq0Beads}
\usage{
addMapq0BeadsRep(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
addMapq0Beads
}
\examples{
#addMapq0Beads('AA001')
}
\author{
Przemyslaw Stempor
}
\seealso{
Other tracks: \code{\link{addMapq0Beads}}
}

