% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakTools.R
\name{addMapq0Track}
\alias{addMapq0Track}
\title{Run peak calls and optionally adds them to DB}
\usage{
addMapq0Track(ids, local = FALSE)
}
\arguments{
\item{local}{should the job be run on local or remote FS}

\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
Run peak calls and optionally adds them to DB
}
\examples{
#addMapq0Track('AA001')
}
\author{
Przemyslaw Stempor
}
\seealso{
Other Peaks: \code{\link{callPeaksMACS}},
  \code{\link{combinePeaksToBed}},
  \code{\link{combinePeaks}},
  \code{\link{enrichedRegionsCall}}
}

