% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTracksTools.R
\name{addMapq0TrackZcs}
\alias{addMapq0TrackZcs}
\title{addMapq0Track and z-score it}
\usage{
addMapq0TrackZcs(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
addMapq0Track and z-score it
}
\examples{
#addMapq0Track('AA001')
}
\seealso{
Other Peaks: \code{\link{addBEADSmapq0TrackZcs}},
  \code{\link{addMapq0Track}},
  \code{\link{addStrandedRNAseq}},
  \code{\link{callPeaksMACS}},
  \code{\link{combinePeaksToBed}},
  \code{\link{combinePeaks}},
  \code{\link{enrichedRegionsCall}},
  \code{\link{jadb_addScaledTrack}},
  \code{\link{jadb_addTracksFromBAM}}
}
\author{
Przemyslaw Stempor
}
