% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jadb_beads.R
\name{addNonUniqueNQBeads}
\alias{addNonUniqueNQBeads}
\title{addNonUniqueQ10Beads}
\usage{
addNonUniqueNQBeads(ids, genome = "ce11")
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
addNonUniqueQ10Beads
}
\examples{
#addNonUniqueQ10Beads('AA001')
}
\seealso{
Other tracks: \code{\link{addMapq0BeadsRep}},
  \code{\link{addMapq0Beads}},
  \code{\link{addNonUniqueQ10Beads}},
  \code{\link{addUniqueQ10Beads}},
  \code{\link{jadb_addAlignedBAM}}
}
\author{
Przemyslaw Stempor
}
