% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTracksTools.R
\name{addNonUniqueQ10Beads}
\alias{addNonUniqueQ10Beads}
\title{addNonUniqueQ10Beads}
\usage{
addNonUniqueQ10Beads(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
addNonUniqueQ10Beads
}
\examples{
#addMapq0Beads('AA001')
}
\seealso{
Other tracks: \code{\link{addMapq0BeadsRep}},
  \code{\link{addMapq0Beads}}
}
\author{
Przemyslaw Stempor
}
