% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jadb_beads.R
\name{addUniqueQ10Beads}
\alias{addUniqueQ10Beads}
\title{addUniqueQ10Beads}
\usage{
addUniqueQ10Beads(ids, genome = "ce11")
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
addUniqueQ10Beads
}
\examples{
#addNonUniqueQ10Beads('AA001')
}
\seealso{
Other tracks: \code{\link{addMapq0BeadsRep}},
  \code{\link{addMapq0Beads}},
  \code{\link{addNonUniqueNQBeads}},
  \code{\link{addNonUniqueQ10Beads}},
  \code{\link{jadb_addAlignedBAM}}
}
\author{
Przemyslaw Stempor
}
