% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jadb_anno.R
\name{anno_add}
\alias{anno_add}
\title{Add annotation to JADB}
\usage{
anno_add(file, Type = "Genes", Version = "ensembl90",
  Description = "tets add", Source = "", Genome = "ce11",
  user = "ps562")
}
\arguments{
\item{file}{- path to file or GR}

\item{Type}{Type}

\item{Version}{Version}

\item{Description}{Description}

\item{Source}{Source}

\item{Genome}{Genome}

\item{user}{user ID}
}
\description{
Add annotation to JADB
}
