% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jadb_anno.R
\name{anno_add_file}
\alias{anno_add_file}
\title{Title}
\usage{
anno_add_file(ID, file, ref.genome.version = "ce11")
}
\arguments{
\item{ID}{ContactExpID}

\item{file}{path to file}

\item{ref.genome.version}{genome version}
}
\description{
Title
}
