% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BAMtools.R
\name{bam2del}
\alias{bam2del}
\title{Calculates delation candidates and outputs them as BAM}
\usage{
bam2del(ID, minsize = 100L, download = TRUE)
}
\arguments{
\item{ID}{the vector of ContactExpIDs}

\item{minsize}{minimum size of delation}

\item{download}{decides if file will be downloaded to local file system}
}
\value{
GRanges with delation candidats
}
\description{
Calculates delation candidates and outputs them as BAM
}
\examples{
#d1 <- bam2del('RC025')
#d2 <- bam2del('RC037')
#export.bed(intersect(d1, d2), 'RC025andRC037_del_more100)intersect.bed')
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{doDiffExpr}};
  \code{\link{downloadSummarizedEperiment}};
  \code{\link{getAnno}};
  \code{\link{getDEreportForDESeqDataSet}};
  \code{\link{getDEreport}}; \code{\link{getDEseq2res}};
  \code{\link{getStage}}; \code{\link{getStrain}};
  \code{\link{getSummarizedEperiment}};
  \code{\link{ids2SE}}; \code{\link{ids2diffExpr}};
  \code{\link{summarizeBAMs}}; \code{\link{wigs2SE}}
}

