% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakTools.R
\name{callPeaksMACS}
\alias{callPeaksMACS}
\title{Run peak calls and optionally adds them to DB}
\usage{
callPeaksMACS(ids, local = TRUE, extsize = 150, summedinput = TRUE)
}
\arguments{
\item{local}{should the job be run on local or remote FS}

\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
Run peak calls and optionally adds them to DB
}
\examples{
#callPeaksMACS(IDs)
}
\author{
Przemyslaw Stempor
}
\seealso{
Other Peaks: \code{\link{addBEADSmapq0TrackZcs}},
  \code{\link{addMapq0TrackZcs}},
  \code{\link{addMapq0Track}},
  \code{\link{addStrandedRNAseq}},
  \code{\link{combinePeaksToBed}},
  \code{\link{combinePeaks}},
  \code{\link{enrichedRegionsCall}}
}

