% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakTools.R
\name{combinePeaks}
\alias{combinePeaks}
\title{Combines peak in replicates}
\usage{
combinePeaks(ids, mode = "u")
}
\arguments{
\item{mode}{u for union or i for intersection}

\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
GRanges
}
\description{
Combines peak in replicates
}
\examples{
#combinePeaks(IDs)
}
\author{
Przemyslaw Stempor
}
\seealso{
Other Peaks: \code{\link{addMapq0TrackZcs}},
  \code{\link{addMapq0Track}},
  \code{\link{addStrandedRNAseq}},
  \code{\link{callPeaksMACS}},
  \code{\link{combinePeaksToBed}},
  \code{\link{enrichedRegionsCall}}
}

