% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakTools.R
\name{combinePeaksToBed}
\alias{combinePeaksToBed}
\title{Combines peak in replicates and exports to bed file}
\usage{
combinePeaksToBed(ids, mode = "union")
}
\arguments{
\item{mode}{union or intersection}

\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
bed file path
}
\description{
Combines peak in replicates and exports to bed file
}
\examples{
#combinePeaksToBed(IDs)
}
\author{
Przemyslaw Stempor
}
\seealso{
Other Peaks: \code{\link{callPeaksMACS}},
  \code{\link{combinePeaks}}
}

