% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BWtools.R
\name{combineReps}
\alias{combineReps}
\title{Combine big wiggle files in the database into replicate track}
\usage{
combineReps(IDs, processing = "aligned", res = 100L, outdir = tempdir())
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}

\item{processing}{Processing level of tracks}

\item{res}{Resolution of tracks}

\item{outdir}{Directory to output BW file}
}
\value{
List
}
\description{
Combine big wiggle files in the database into replicate track
}
\examples{
#combineReps(IDs)
}
\author{
Przemyslaw Stempor
}
\seealso{
Other BW: \code{\link{addRepToJADB}}
}

