% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_rnaseq.R
\name{db_mran_rpm_track}
\alias{db_mran_rpm_track}
\title{db_mran_rpm_track}
\usage{
db_mran_rpm_track(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
null
}
\description{
db_mran_rpm_track
}
\seealso{
Other db_rnaseq: \code{\link{db_mran_seq_align_process}},
  \code{\link{jadb_align_rnaseq}},
  \code{\link{jadb_trimm_fq}}
}
\author{
Przemyslaw Stempor
}
