% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_basespace.R
\name{download_fastq_from_basespace}
\alias{download_fastq_from_basespace}
\title{download_fastq_from_basespace}
\usage{
download_fastq_from_basespace(data)
}
\arguments{
\item{data}{}
}
\value{
file path
}
\description{
download_fastq_from_basespace
}
