% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakTools.R
\name{enrichedRegionsCall}
\alias{enrichedRegionsCall}
\title{Calls enriched regions from BigWig file}
\usage{
enrichedRegionsCall(file, bedoutput = NULL)
}
\arguments{
\item{file}{path/url to BW}

\item{bedoutput}{path to output BED}
}
\value{
GRanges or bed file path
}
\description{
Calls enriched regions from BigWig file
}
\examples{
#combinePeaksToBed(IDs)
}
\seealso{
Other Peaks: \code{\link{addBEADSmapq0TrackZcs}},
  \code{\link{addMapq0TrackZcs}},
  \code{\link{addMapq0Track}},
  \code{\link{addStrandedRNAseq}},
  \code{\link{callPeaksMACS}},
  \code{\link{combinePeaksToBed}},
  \code{\link{combinePeaks}},
  \code{\link{jadb_addScaledTrack}},
  \code{\link{jadb_addTracksFromBAM}}
}
\author{
Przemyslaw Stempor
}
