% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_utils.R
\name{fiXPath}
\alias{fiXPath}
\title{Fixes names}
\usage{
fiXPath(ID)
}
\arguments{
\item{ID}{ContactExpID of fixed entry}
}
\description{
Fixes names
}
\examples{
#sapply(sprintf('rML\%.3i', 3:24) , fiXPath)

}
\author{
Przemyslaw Stempor
}
\seealso{
Other fix: \code{\link{fiXWig}}
}

