% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_utils.R
\name{fiXWig}
\alias{fiXWig}
\title{Fixes overlaps in bed files}
\usage{
fiXWig(x, gnm = BSgenome.Celegans.UCSC.ce10::Celegans, pth = sub("wig",
  "bw", x))
}
\arguments{
\item{x}{path to bw}

\item{gnm}{BSgenome}

\item{pth}{output path}
}
\description{
Fixes overlaps in bed files
}
\examples{
#sapply(dir(pattern='wig') , fiXWig)

}
\seealso{
Other fix: \code{\link{fiXPath}}
}
\author{
Przemyslaw Stempor
}
