% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_add_fq.R
\name{fs_add_internal}
\alias{fs_add_internal}
\title{Add files from internal FS}
\usage{
fs_add_internal(addr, select_id = "", EXTABLE = "labexperiment",
  ignore.exist = FALSE)
}
\arguments{
\item{addr}{}

\item{select_id}{}

\item{EXTABLE}{}

\item{ignore.exist}{}
}
\value{
res object
}
\description{
Add files from internal FS
}
