% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_add_fq.R
\name{fs_insert_fastq_to_jadb}
\alias{fs_insert_fastq_to_jadb}
\title{Add fastq to existing experiment}
\usage{
fs_insert_fastq_to_jadb(id, finalFilePath, EXTABLE = "labrnaseq")
}
\arguments{
\item{id}{}

\item{finalFilePath}{}

\item{EXTABLE}{}
}
\description{
Add fastq to existing experiment
}
