% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEOtoBW.R
\name{GEOarray2bw}
\alias{GEOarray2bw}
\title{Get BigWig tracks from GEO microarray experiment}
\usage{
GEOarray2bw(ID, platform = "GPL8647", liftover = TRUE)
}
\arguments{
\item{ID}{list of GSM IDs}

\item{platform}{GPL id of platform}

\item{liftover}{if TRUE lifts over from ce4 tp ce10}
}
\value{
List of tracks as RleList
}
\description{
Get BigWig tracks from GEO microarray experiment
}
\examples{
input <- list(
    "H3K36me3_mes-4 RNAi_rep1"="GSM938394",
    "H3K36me3_mes-4 RNAi_rep2"="GSM938395",
    "H3K27me3_mes-4 RNAi_rep1"="GSM938396",
    "H3K27me3_mes-4 RNAi_rep2"="GSM938397",
    "H3K36me3_N2_rep1"="GSM938401",
    "H3K36me3_N2_rep2"="GSM938402",
    "H3K36me3_N2_rep3"="GSM938403",
    "H3K27me3_N2_rep1"="GSM938404",
    "H3K27me3_N2_rep2"="GSM938405"
)

ANS <- GEOarray2bw(input)
H3K36me3_mes4 <- (ANS$"H3K36me3_mes-4 RNAi_rep1" + ANS$"H3K36me3_mes-4 RNAi_rep2")/2
export.bw(H3K36me3_mes4 , 'H3K36me3_mes4.bw')

H3K27me3_mes4 <- (ANS$"H3K27me3_mes-4 RNAi_rep1" + ANS$"H3K27me3_mes-4 RNAi_rep2")/2
export.bw(H3K27me3_mes4, 'H3K27me3_mes4.bw')

H3K36me3_N2 <- (ANS$"H3K36me3_N2_rep1" + ANS$"H3K36me3_N2_rep2" + ANS$"H3K36me3_N2_rep3")/3
export.bw(H3K36me3_N2, 'H3K36me3_N2.bw')

H3K27me3_N2    <- (ANS$"H3K27me3_N2_rep1"+ ANS$"H3K27me3_N2_rep2")/2
export.bw(H3K27me3_N2, 'H3K27me3_N2.bw')
}
\author{
Przemyslaw Stempor
}

