% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shortRNA.R
\name{get21Uand22G}
\alias{get21Uand22G}
\title{Extract 21U and 22G short RNA specied from FASTQ file}
\usage{
get21Uand22G(file)
}
\arguments{
\item{file}{the path to input FASTTQ}
}
\value{
NULL; creates 2 FASTQ file in the dirname of file
}
\description{
Extract 21U and 22G short RNA specied from FASTQ file
}
\author{
Przemyslaw Stempor
}

