% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortRNA.R
\name{get21Uand22GfromIDs}
\alias{get21Uand22GfromIDs}
\title{Extract 21U and 22G short RNA specied from FASTQ file}
\usage{
get21Uand22GfromIDs(IDs)
}
\arguments{
\item{IDs}{IDs of trimmed files}
}
\value{
NULL; creates 2 FASTQ file in the dirname of file
}
\description{
Extract 21U and 22G short RNA specied from FASTQ file
}
\seealso{
Other shortRNA: \code{\link{get21Uand22G}}
}
\author{
Przemyslaw Stempor
}
