% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{getDBdataField}
\alias{getDBdataField}
\title{Get data filed from IDs}
\usage{
getDBdataField(ContactExpID, field = "ContactExpID", EXTABLE = "labrnaseq")
}
\arguments{
\item{ContactExpID}{vector of IDs}

\item{field}{name of the nannotation field}
}
\value{
eset
}
\description{
Get data filed from IDs
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{bam2del}};
  \code{\link{doDiffExpr}};
  \code{\link{downloadSummarizedEperiment}};
  \code{\link{getAnno}};
  \code{\link{getDEreportForDESeqDataSet}};
  \code{\link{getDEreport}}; \code{\link{getDEseq2res}};
  \code{\link{getStage}}; \code{\link{getStrain}};
  \code{\link{getSummarizedEperiment}};
  \code{\link{ids2SE}}; \code{\link{ids2diffExpr}};
  \code{\link{spikeINsummary}};
  \code{\link{summarizeBAMs}}; \code{\link{wigs2SE}}
}

