% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{getDEreport}
\alias{getDEreport}
\title{Get DEseq2 report as HTML}
\usage{
getDEreport(ID, path = NULL, location = "DEreport", show = TRUE)
}
\arguments{
\item{ID}{id}
}
\value{
file colation
}
\description{
Get DEseq2 report as HTML
}
\seealso{
Other RNAseq: \code{\link{bam2del}},
  \code{\link{doDiffExpr}},
  \code{\link{downloadSummarizedEperiment}},
  \code{\link{getAnno}}, \code{\link{getDBdataField}},
  \code{\link{getDEreportForDESeqDataSet}},
  \code{\link{getDEseq2res}}, \code{\link{getStage}},
  \code{\link{getStrain}},
  \code{\link{getSummarizedEperiment}},
  \code{\link{ids2SE}}, \code{\link{ids2diffExpr}},
  \code{\link{spikeINsummary}},
  \code{\link{summarizeBAMs}}, \code{\link{wigs2SE}}
}
\author{
Przemyslaw Stempor
}
