% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDEreportForDESeqDataSet}
\alias{getDEreportForDESeqDataSet}
\title{Get DEseq2 report as HTML}
\usage{
getDEreportForDESeqDataSet(dds, location = "DEreport", show = TRUE)
}
\arguments{
\item{dds}{DESeqDataSet form DESeq2 package}

\item{location}{sub-directory to create the report}

\item{show}{should R open the report in the browser}
}
\value{
file colation
}
\description{
Get DEseq2 report as HTML
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{doDiffExpr}};
  \code{\link{getDEreport}}; \code{\link{getDEseq2res}};
  \code{\link{getStrain}}; \code{\link{ids2diffExpr}};
  \code{\link{summarizeBAMs}}
}

