% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{getDEseq2res}
\alias{getDEseq2res}
\title{Get DEseq2 redult from DB}
\usage{
getDEseq2res(ID, path = NULL)
}
\arguments{
\item{ID}{id}

\item{path}{mmt path, return URL if null}
}
\value{
list
}
\description{
Get DEseq2 redult from DB
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{bam2del}};
  \code{\link{doDiffExpr}};
  \code{\link{getDEreportForDESeqDataSet}};
  \code{\link{getDEreport}}; \code{\link{getStrain}};
  \code{\link{ids2diffExpr}}; \code{\link{summarizeBAMs}}
}

