% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JADBtools.R
\name{getFilePath}
\alias{getFilePath}
\title{Get file URL}
\usage{
getFilePath(ID, format = ".", processing = ".", scale = ".", url = TRUE,
  eq = FALSE)
}
\arguments{
\item{ID}{ContactExperimetID}

\item{format}{filter regex}

\item{processing}{filter regex}

\item{scale}{filter regex}

\item{url}{return url or just the path to the file}

\item{eq}{t}
}
\value{
URL string
}
\description{
Accepts regular expressions.
}
\author{
Przemyslaw Stempor
}
\seealso{
Other dbtools: \code{\link{addGenericFile}};
  \code{\link{formGenericPath}}; \code{\link{getFileUID}};
  \code{\link{makeFileUID}}
}

