% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikeInNorm.R
\name{getRatio}
\alias{getRatio}
\title{Get ratio between worm and 10x fly reads}
\usage{
getRatio(ID)
}
\arguments{
\item{ID}{the database id of the experiment in "XX001" format}
}
\value{
named \code{\link[base]{numeric}} value
}
\description{
This ratio can be used as global normalizing parameter for sipke in.
}
\seealso{
Other spikeIn: \code{\link{normTrack}}
}
\author{
Przemyslaw Stempor
}
