% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idr_functions.R
\name{idr_concave_on_replicate}
\alias{idr_concave_on_replicate}
\title{Single-pass IDR from replicate}
\usage{
idr_concave_on_replicate(id, processing = "BEADSQ10NU",
  curvature_idx = 500L)
}
\arguments{
\item{id}{ID of second experiment in replicate}

\item{processing}{}

\item{curvature_idx}{}
}
\value{
GRanges
}
\description{
Single-pass IDR from replicate
}
