% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idr_functions.R
\name{idr_concave_two_pass_with_oracle}
\alias{idr_concave_two_pass_with_oracle}
\title{Two-pass IDR procedure with oracle peak call set}
\usage{
idr_concave_two_pass_with_oracle(r1, r2, processing = "BEADSQ10NU",
  curvature_idx = 100)
}
\arguments{
\item{r1}{ID of first experiment in replicate}

\item{r2}{ID of second experiment in replicate}

\item{processing}{Which tracks to use (default: BEADSQ10NU)}

\item{curvature_idx}{- curvature index treshold (default: 100)}
}
\description{
Two-pass IDR procedure with oracle peak call set
}
