% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ids2diffExpr}
\alias{ids2diffExpr}
\title{Run full differential expression pipeline}
\usage{
ids2diffExpr(ID, mnt_point = NULL)
}
\arguments{
\item{ID}{the vector of ContactExpIDs}

\item{mnt_point}{give the mount point for filesystem, use URLs if NULL, defaults to NULL}
}
\value{
stats string
}
\description{
Run full differential expression pipeline
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{doDiffExpr}};
  \code{\link{getStrain}}; \code{\link{summarizeBAMs}}
}

