% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DifferentialPantterns.R
\name{ids2SE}
\alias{ids2SE}
\title{Run full differential expression pipeline}
\usage{
ids2SE(ID, genes = unlist(range(get(data(gnmodel)))), mnt_point = NULL)
}
\arguments{
\item{ID}{the vector of ContactExpIDs}

\item{mnt_point}{give the mount point for filesystem, use URLs if NULL, defaults to NULL}
}
\value{
stats string
}
\description{
Run full differential expression pipeline
}
\examples{
ID=c("AM004","FB006","FB013merge","FB008","FB014merge","FB009","FB015merge","AA103p","AA076","AA186","AA378","BC137","BC131","AA077","AA078","BC138","BC132","AA089","AA079")
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{bam2del}};
  \code{\link{doDiffExpr}};
  \code{\link{downloadSummarizedEperiment}};
  \code{\link{getAnno}};
  \code{\link{getDEreportForDESeqDataSet}};
  \code{\link{getDEreport}}; \code{\link{getDEseq2res}};
  \code{\link{getStage}}; \code{\link{getStrain}};
  \code{\link{getSummarizedEperiment}};
  \code{\link{ids2diffExpr}}; \code{\link{summarizeBAMs}};
  \code{\link{wigs2SE}}
}

