% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_basespace.R
\name{insert_entry_to_jadb}
\alias{insert_entry_to_jadb}
\title{insert_entry_to_jadb}
\usage{
insert_entry_to_jadb(data, finalFilePath)
}
\arguments{
\item{data}{}

\item{finalFilePath}{}
}
\description{
insert_entry_to_jadb
}
