% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jadb_cluster_ops.R
\name{jacl_mass_parallel}
\alias{jacl_mass_parallel}
\title{jacl_mass_parallel}
\usage{
jacl_mass_parallel(n = 50)
}
\arguments{
\item{ContactExpID}{vector of IDs}
}
\value{
eset
}
\description{
jacl_mass_parallel
}
\seealso{
Other cluster: \code{\link{jacl_send_to_cluster_ce10}},
  \code{\link{jacl_send_to_cluster}}
}
\author{
Przemyslaw Stempor
}
