% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jadb_cluster_ops.R
\name{jacl_send_to_cluster_ce10}
\alias{jacl_send_to_cluster_ce10}
\title{jacl_send_to_cluster_ce10 for ce10}
\usage{
jacl_send_to_cluster_ce10(ID)
}
\arguments{
\item{ID}{}
}
\description{
jacl_send_to_cluster_ce10 for ce10
}
\examples{
# sapply(ids, jacl_send_to_cluster_ce10)
}
\seealso{
Other cluster: \code{\link{jacl_mass_parallel}},
  \code{\link{jacl_send_to_cluster}}
}
\author{
Przemyslaw Stempor
}
