% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_alignment.R
\name{jadb_addAlignedBAM}
\alias{jadb_addAlignedBAM}
\title{jadb_addAlignedBAM}
\usage{
jadb_addAlignedBAM(ids, processing = "raw", format = "txt.gz",
  genome = "ce11")
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
jadb_addAlignedBAM
}
\examples{
#jadb_addAlignedBAM('AA001')
#parallel::mclapply(sprintf('REP\%.3i', 36:42)[-6], jadb_addAlignedBAM), mc.cores = 8)
}
\seealso{
Other tracks: \code{\link{addMapq0BeadsRep}},
  \code{\link{addMapq0Beads}},
  \code{\link{addNonUniqueNQBeads}},
  \code{\link{addNonUniqueQ10Beads}}
}
\author{
Przemyslaw Stempor
}
