% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_trackops.R
\name{jadb_addTracksFromBAM}
\alias{jadb_addTracksFromBAM}
\title{jadb_addTracksFromBAM}
\usage{
jadb_addTracksFromBAM(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
jadb_addTracksFromBAM
}
\examples{
#jadb_addTracksFromBAM('AA001')
}
\seealso{
Other Peaks: \code{\link{addBEADSmapq0TrackZcs}},
  \code{\link{addMapq0TrackZcs}},
  \code{\link{addMapq0Track}},
  \code{\link{addStrandedRNAseq}},
  \code{\link{callPeaksMACS}},
  \code{\link{combinePeaksToBed}},
  \code{\link{combinePeaks}},
  \code{\link{enrichedRegionsCall}}
}
\author{
Przemyslaw Stempor
}
