% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_rnaseq.R
\name{jadb_align_rnaseq}
\alias{jadb_align_rnaseq}
\title{jadb_align_rnaseq}
\usage{
jadb_align_rnaseq(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
null
}
\description{
jadb_align_rnaseq
}
\seealso{
Other db_rnaseq: \code{\link{db_mran_rpm_track}},
  \code{\link{db_mran_seq_align_process}},
  \code{\link{jadb_trimm_fq}}
}
\author{
Przemyslaw Stempor
}
