% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_basespace.R
\name{jadb_basespace}
\alias{jadb_basespace}
\title{Title}
\usage{
jadb_basespace(addr, EXTABLE = "mydb.labexperiment", ignore.exist = FALSE,
  select_id = "")
}
\arguments{
\item{addr}{}

\item{EXTABLE}{}

\item{ignore.exist}{}

\item{select_id}{}
}
\value{
result
}
\description{
Title
}
