% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jadb_cluster_ops.R
\name{jadb_dc}
\alias{jadb_dc}
\title{jadb_dc - direct submission to JADB}
\usage{
jadb_dc(gurl, genome = c("ce11", "cb3ce11"), legacy_ce10 = TRUE,
  wipeout = FALSE)
}
\arguments{
\item{gurl}{- Google Docs URL}

\item{genome}{- reference genome}

\item{legacy_ce10}{- add entries to ce10 database}
}
\description{
jadb_dc - direct submission to JADB
}
