% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_access_functions.R
\name{jadb_download_files}
\alias{jadb_download_files}
\title{jadb_download_files}
\usage{
jadb_download_files(ids, processing = "raw", auto = FALSE)
}
\arguments{
\item{ids}{ContactExpIDs}

\item{processing}{raw}
}
\value{
TRUE if downloaded
}
\description{
jadb_download_files
}
