% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_access_functions.R
\name{jadb_get_file_paths}
\alias{jadb_get_file_paths}
\title{jadb_get_file_paths}
\usage{
jadb_get_file_paths(ids, processing = "raw",
  root_prefix = "/mnt/jadb/DBfile/DBfiles")
}
\arguments{
\item{ids}{ContactExpIDs}

\item{processing}{raw}

\item{root_prefix}{/mnt/jadb/DBfile/DBfiles}
}
\value{
paths
}
\description{
jadb_get_file_paths
}
