% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_alignment.R
\name{jadb_processs_sceleton}
\alias{jadb_processs_sceleton}
\title{jadb_processs_sceleton}
\usage{
jadb_processs_sceleton(ids, FUN, Processing = "TETS", Scale = "NA",
  Resolution = "NA", filetype_format = "unknown", format = "bam",
  processing = "aligned", scale = "NA", resolution = "NA", ...)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
jadb_processs_sceleton
}
\examples{
#jadb_processs_sceleton(
#    ids, FUN = run_fastqc, 
#    Processing = 'FastQC', Scale = 'NA', Resolution = 'NA', filetype_format = 'HTML',
#    format = "bam", processing = "aligned", scale = "NA", resolution='NA'
#)
}
\author{
Przemyslaw Stempor
}
