% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSops.R
\name{jadb_purge_exp}
\alias{jadb_purge_exp}
\title{Purge an experiment from all processed files}
\usage{
jadb_purge_exp(ids)
}
\arguments{
\item{ids}{}
}
\description{
Purge an experiment from all processed files
}
