% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_rename_files.R
\name{jadb_rename_chip}
\alias{jadb_rename_chip}
\title{Rename ChIP experiments}
\usage{
jadb_rename_chip(ID)
}
\arguments{
\item{ID}{- ContactExpID of an experiment}
}
\description{
Rename ChIP experiments
}
