% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{jadb_replicates_chip}
\alias{jadb_replicates_chip}
\title{Adds replicates to database}
\usage{
jadb_replicates_chip(IDs, res = 100L, prefix = "REP", extra_stats = NULL)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}

\item{res}{Resolution of tracks}

\item{outdir}{Directory to output BW file}
}
\description{
Adds replicates to database
}
\examples{
#combineReps(IDs)
}
\seealso{
Other BW: \code{\link{addIDRtoJADB}},
  \code{\link{combineReps}}
}
\author{
Przemyslaw Stempor
}
