% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_pipelenies.R
\name{jadb_RNAseq}
\alias{jadb_RNAseq}
\title{jadb_RNAseq}
\usage{
jadb_RNAseq(ids, steps = c("trim", "aln", "tracks", "rpm", "fastqc", "star"),
  ...)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
jadb_RNAseq
}
\examples{
#
}
\seealso{
Other pipelines: \code{\link{jadb_ChIPseq}}
}
\author{
Przemyslaw Stempor
}
