% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_rnaseq.R
\name{jadb_trimm_fq}
\alias{jadb_trimm_fq}
\title{jadb_trimm_fq}
\usage{
jadb_trimm_fq(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
null
}
\description{
jadb_trimm_fq
}
\seealso{
Other db_rnaseq: \code{\link{db_mran_rpm_track}},
  \code{\link{db_mran_seq_align_process}},
  \code{\link{jadb_align_rnaseq}}
}
\author{
Przemyslaw Stempor
}
