% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JADBtools-package.R
\docType{package}
\name{JADBtools}
\alias{JADBtools}
\alias{JADBtools-package}
\title{JADB I/O tools}
\description{
Tools for retrive and manipulate data in JADB.
}
\author{
Przemyslaw Stempor
}

