% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_basespace.R
\name{jagui_add_submission_csv}
\alias{jagui_add_submission_csv}
\title{jagui_add_submission_csv}
\usage{
jagui_add_submission_csv(gurl, verbose = FALSE, host = NULL)
}
\arguments{
\item{gurl}{url}

\item{verbose}{hots}

\item{host}{show commands}
}
\value{
SQL DBinfo
}
\description{
jagui_add_submission_csv
}
