% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_basespace.R
\name{jagui_get_table}
\alias{jagui_get_table}
\title{jagui_get_table}
\usage{
jagui_get_table(tab)
}
\arguments{
\item{tab}{name of MySQL table, e.g. 'labexperiment'}
}
\value{
tibble
}
\description{
jagui_get_table
}
\examples{
# jagui_get_table('labexperiment')
}
