% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_remote.R
\name{jar_log}
\alias{jar_log}
\title{Title}
\usage{
jar_log(p = ".chip", n = 3, db = "jadb")
}
\arguments{
\item{p}{- pattern to match}

\item{n}{- number of linest to print}

\item{db}{- database to search}
}
\description{
Title
}
