% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_remote.R
\name{jar_nlog}
\alias{jar_nlog}
\title{jar_nlog}
\usage{
jar_nlog(p = ".chip", n = 3, db = "jadb",
  remote = "jarun@cb-head2.gurdon.private.cam.ac.uk")
}
\arguments{
\item{p}{- pattern to match}

\item{n}{- number of linest to print}

\item{db}{- database to search}
}
\description{
jar_nlog
}
