% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_remote.R
\name{jar_process_exp}
\alias{jar_process_exp}
\title{jar_process_exp}
\usage{
jar_process_exp(IDS, genome = "ce11")
}
\arguments{
\item{IDS}{- ContactExpIDs}

\item{genome}{- Genome version}
}
\description{
jar_process_exp
}
\examples{
#jar_process_exp( JADBtools:::make_jadb_ids(106:112, 'CG'), genome = 'cb3ce11' )

}
