% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_access_functions.R
\name{make_jadb_ids}
\alias{make_jadb_ids}
\title{make_jadb_ids}
\usage{
make_jadb_ids(ran, prefix = "rYD")
}
\arguments{
\item{ran}{range}

\item{prefix}{prefix}
}
\value{
ids
}
\description{
make_jadb_ids
}
