% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeFileUID}
\alias{makeFileUID}
\title{Make UID}
\usage{
makeFileUID(string = "", prefix = "X")
}
\arguments{
\item{string}{file patch or other string for md5}

\item{prefix}{UID prefix letter, defaults to "X"}
}
\value{
IUD string
}
\description{
Make UID
}
\author{
Przemyslaw Stempor
}
\seealso{
Other dbtools: \code{\link{addGenericFile}};
  \code{\link{formGenericPath}}; \code{\link{getFilePath}}
}

