% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_motif_tools.R
\name{meme_chip_local}
\alias{meme_chip_local}
\title{meme_chip_local}
\usage{
meme_chip_local(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
meme_chip_local
}
\examples{
#meme_chip_local('AA001')
}
\seealso{
Other Peaks: \code{\link{addBEADSmapq0TrackZcs}},
  \code{\link{addMapq0TrackZcs}},
  \code{\link{addMapq0Track}},
  \code{\link{addStrandedRNAseq}},
  \code{\link{callPeaksMACS}},
  \code{\link{combinePeaksIDR}},
  \code{\link{combinePeaksToBed}},
  \code{\link{combinePeaks}},
  \code{\link{enrichedRegionsCall}},
  \code{\link{jadb_addScaledTrack}},
  \code{\link{jadb_addTracksFromBAM}}
}
\author{
Przemyslaw Stempor
}
