% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idr_functions.R
\name{overlap_intersect}
\alias{overlap_intersect}
\title{Overlap intersect of two peak calls}
\usage{
overlap_intersect(r1, r2, processing = "BEADSQ10NU", dc = TRUE)
}
\arguments{
\item{r1}{ID of first experiment in replicate}

\item{r2}{ID of second experiment in replicate}

\item{processing}{Which tracks to use (default: BEADSQ10NU)}

\item{dc}{Create new directory}
}
\description{
Overlap intersect of two peak calls
}
