% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{pb}
\alias{pb}
\title{Print progress bar JSON}
\usage{
pb(mnt, uid, s = TRUE, p = 0.01, t = "Loading libraries...", tt = "")
}
\value{
message or error
}
\description{
Print progress bar JSON
}
\author{
Przemyslaw Stempor
}
\seealso{
Other dbutils: \code{\link{testConnection}}
}

