% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raporting_tools.R
\name{renderReport}
\alias{renderReport}
\title{ChIP profiles enrichment of repetitive elements in C. elegans raport}
\usage{
renderReport(infile)
}
\arguments{
\item{infile}{BigWig file path}
}
\value{
HTML file
}
\description{
ChIP profiles enrichment of repetitive elements in C. elegans raport
}
\author{
Przemyslaw Stempor
}

