% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testConnection}
\alias{testConnection}
\title{Test connection to JADB}
\usage{
testConnection()
}
\value{
message or error
}
\description{
Test connection to JADB
}
\author{
Przemyslaw Stempor
}
\seealso{
Other dbutils: \code{\link{pb}}
}

