% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_basespace.R
\name{validate_jadb_submission_entry}
\alias{validate_jadb_submission_entry}
\title{validate_jadb_submission_entry}
\usage{
validate_jadb_submission_entry(entry, EXTABLE = "mydb.labexperiment",
  ignore.exist = FALSE)
}
\arguments{
\item{entry}{}

\item{EXTABLE}{}

\item{ignore.exist}{}
}
\value{
data structuren (list)
}
\description{
validate_jadb_submission_entry
}
