% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addGenericFile}
\alias{addGenericFile}
\title{Add file to database}
\usage{
addGenericFile(ContactExpID, path, Processing = "raw", Resolution = NA,
  Scale = NA, genome = NA, prefix = "X", filetype_format = "txt.gz",
  comments = "", userID = NA, parent1_uid = NA, parent2_uid = NA,
  parent3_uid = NA, uniq = NA)
}
\arguments{
\item{ContactExpID}{ID string}

\item{path}{prefix UID prefix letter, defaults to "X"}
}
\value{
entry list
}
\description{
Add file to database
}
\author{
Przemyslaw Stempor
}
\seealso{
Other dbtools: \code{\link{formGenericPath}};
  \code{\link{getFilePath}}; \code{\link{makeFileUID}}
}

