% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWtools.R
\name{addIDRtoJADB}
\alias{addIDRtoJADB}
\title{OLD IDR reps}
\usage{
addIDRtoJADB(IDs, res = 100L)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}

\item{res}{Resolution of tracks}

\item{outdir}{Directory to output BW file}
}
\description{
OLD IDR reps
}
\examples{
#combineReps(IDs)
}
\seealso{
Other BW: \code{\link{combineReps}},
  \code{\link{jadb_replicates_chip}}
}
\author{
Przemyslaw Stempor
}
