% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTracksTools.R
\name{addMapq0BeadsRep}
\alias{addMapq0BeadsRep}
\title{addMapq0BeadsRep}
\usage{
addMapq0BeadsRep(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
addMapq0BeadsRep
}
\examples{
#addMapq0Beads('AA001')
}
\seealso{
Other tracks: \code{\link{addMapq0Beads}},
  \code{\link{addNonUniqueNQBeads}},
  \code{\link{addNonUniqueQ10Beads}},
  \code{\link{addUniqueQ10Beads}},
  \code{\link{jadb_addAlignedBAM}}
}
\author{
Przemyslaw Stempor
}
