% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTracksTools.R
\name{addMapq0Track}
\alias{addMapq0Track}
\title{addMapq0Track}
\usage{
addMapq0Track(ids)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
addMapq0Track
}
\examples{
#addMapq0Track('AA001')
}
\seealso{
Other Peaks: \code{\link{addBEADSmapq0TrackZcs}},
  \code{\link{addMapq0TrackZcs}},
  \code{\link{addStrandedRNAseq}},
  \code{\link{callPeaksMACS}},
  \code{\link{combinePeaksToBed}},
  \code{\link{combinePeaks}},
  \code{\link{enrichedRegionsCall}}
}
\author{
Przemyslaw Stempor
}
