% Generated by roxygen2 (4.0.2): do not edit by hand
\name{doDiffExpr}
\alias{doDiffExpr}
\title{Run DEseq2 and output results}
\usage{
doDiffExpr(e, design = ~strain)
}
\arguments{
\item{e}{eset}

\item{design}{experiment design}
}
\value{
list
}
\description{
Run DEseq2 and output results
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{getDEreportForDESeqDataSet}};
  \code{\link{getDEreport}}; \code{\link{getDEseq2res}};
  \code{\link{getStrain}}; \code{\link{ids2diffExpr}};
  \code{\link{summarizeBAMs}}
}

