% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{downloadSummarizedEperiment}
\alias{downloadSummarizedEperiment}
\title{Get SummarizedEperiments as Rdata from IDs}
\usage{
downloadSummarizedEperiment(ContactExpIDs = "rAM022", dir = "SE")
}
\arguments{
\item{ContactExpID}{vector of IDs}
}
\value{
SummarizedEperiment
}
\description{
Get SummarizedEperiments as Rdata from IDs
}
\seealso{
Other RNAseq: \code{\link{bam2del}},
  \code{\link{doDiffExpr}}, \code{\link{getAnno}},
  \code{\link{getDBdataField}},
  \code{\link{getDEreportForDESeqDataSet}},
  \code{\link{getDEreport}}, \code{\link{getDEseq2res}},
  \code{\link{getStage}}, \code{\link{getStrain}},
  \code{\link{getSummarizedEperiment}},
  \code{\link{ids2SE}}, \code{\link{ids2diffExpr}},
  \code{\link{spikeINsummary}},
  \code{\link{summarizeBAMs}}, \code{\link{wigs2SE}}
}
\author{
Przemyslaw Stempor
}
