% Generated by roxygen2 (4.0.2): do not edit by hand
\name{formGenericPath}
\alias{formGenericPath}
\title{Make generic file name}
\usage{
formGenericPath(ContactExpID, EXTABLE = "labrnaseq", Processing = NA,
  Resolution = NA, Scale = NA)
}
\arguments{
\item{ContactExpID}{experiment ID string}

\item{EXTABLE}{experiment table in db schema}

\item{scale}{file name parameter}

\item{processing}{file name parameter}

\item{format}{file name parameter}
}
\value{
name string
}
\description{
Make generic file name
}
\author{
Przemyslaw Stempor
}
\seealso{
Other dbtools: \code{\link{addGenericFile}};
  \code{\link{getFilePath}}; \code{\link{getFileUID}};
  \code{\link{makeFileUID}}
}

