% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDEreport}
\alias{getDEreport}
\title{Get DEseq2 report as HTML}
\usage{
getDEreport(ID, location = "DEreport", show = TRUE)
}
\arguments{
\item{ID}{id}
}
\value{
file colation
}
\description{
Get DEseq2 report as HTML
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{doDiffExpr}};
  \code{\link{getDEreportForDESeqDataSet}};
  \code{\link{getDEseq2res}}; \code{\link{getStrain}};
  \code{\link{ids2diffExpr}}; \code{\link{summarizeBAMs}}
}

