% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{getDEreportForDESeqDataSet}
\alias{getDEreportForDESeqDataSet}
\title{Get DEseq2 report as HTML}
\usage{
getDEreportForDESeqDataSet(dds, location = "DEreport", show = TRUE)
}
\arguments{
\item{dds}{DESeqDataSet form DESeq2 package}

\item{location}{sub-directory to create the report}

\item{show}{should R open the report in the browser}
}
\value{
file colation
}
\description{
Get DEseq2 report as HTML
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{bam2del}};
  \code{\link{doDiffExpr}};
  \code{\link{downloadSummarizedEperiment}};
  \code{\link{getAnno}}; \code{\link{getDEreport}};
  \code{\link{getDEseq2res}}; \code{\link{getStage}};
  \code{\link{getStrain}};
  \code{\link{getSummarizedEperiment}};
  \code{\link{ids2SE}}; \code{\link{ids2diffExpr}};
  \code{\link{summarizeBAMs}}; \code{\link{wigs2SE}}
}

