% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{getDEseq2res}
\alias{getDEseq2res}
\title{Get DEseq2 redult from DB}
\usage{
getDEseq2res(ID, path = NULL)
}
\arguments{
\item{ID}{id}

\item{path}{mmt path, return URL if null}
}
\value{
list
}
\description{
Get DEseq2 redult from DB
}
\seealso{
Other RNAseq: \code{\link{bam2del}},
  \code{\link{doDiffExpr}},
  \code{\link{downloadSummarizedEperiment}},
  \code{\link{getAnno}}, \code{\link{getDBdataField}},
  \code{\link{getDEreportForDESeqDataSet}},
  \code{\link{getDEreport}}, \code{\link{getStage}},
  \code{\link{getStrain}},
  \code{\link{getSummarizedEperiment}},
  \code{\link{ids2SE}}, \code{\link{ids2diffExpr}},
  \code{\link{spikeINsummary}},
  \code{\link{summarizeBAMs}}, \code{\link{wigs2SE}}
}
\author{
Przemyslaw Stempor
}
