% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFilePath}
\alias{getFilePath}
\title{Get file URL}
\usage{
getFilePath(ID, format = ".", processing = ".", scale = ".")
}
\arguments{
\item{ID}{ContactExperimetID}

\item{format}{filter regex}

\item{processing}{filter regex}

\item{scale}{filter regex}
}
\value{
URL string
}
\description{
Accepts regular expressions.
}
\author{
Przemyslaw Stempor
}
\seealso{
Other dbtools: \code{\link{addGenericFile}};
  \code{\link{formGenericPath}}; \code{\link{makeFileUID}}
}

