% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{getStrain}
\alias{getStrain}
\title{Get strain from IDs}
\usage{
getStrain(ContactExpID, EXTABLE = "labrnaseq")
}
\arguments{
\item{ContactExpID}{vector of IDs}
}
\value{
eset
}
\description{
Get strain from IDs
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{bam2del}};
  \code{\link{doDiffExpr}};
  \code{\link{downloadSummarizedEperiment}};
  \code{\link{getAnno}};
  \code{\link{getDEreportForDESeqDataSet}};
  \code{\link{getDEreport}}; \code{\link{getDEseq2res}};
  \code{\link{getStage}};
  \code{\link{getSummarizedEperiment}};
  \code{\link{ids2SE}}; \code{\link{ids2diffExpr}};
  \code{\link{summarizeBAMs}}; \code{\link{wigs2SE}}
}

