% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getStrain}
\alias{getStrain}
\title{Get starnd from IDs}
\usage{
getStrain(ContactExpID, EXTABLE = "labrnaseq")
}
\arguments{
\item{ContactExpID}{vector of IDs}
}
\value{
eset
}
\description{
Get starnd from IDs
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{doDiffExpr}};
  \code{\link{ids2diffExpr}}; \code{\link{summarizeBAMs}}
}

