% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{getSummarizedEperiment}
\alias{getSummarizedEperiment}
\title{Get SummarizedEperiments from IDs}
\usage{
getSummarizedEperiment(ContactExpIDs = "rAM022")
}
\arguments{
\item{ContactExpID}{vector of IDs}
}
\value{
SummarizedEperiment
}
\description{
Get SummarizedEperiments from IDs
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{bam2del}};
  \code{\link{doDiffExpr}};
  \code{\link{downloadSummarizedEperiment}};
  \code{\link{getAnno}};
  \code{\link{getDEreportForDESeqDataSet}};
  \code{\link{getDEreport}}; \code{\link{getDEseq2res}};
  \code{\link{getStage}}; \code{\link{getStrain}};
  \code{\link{ids2SE}}; \code{\link{ids2diffExpr}};
  \code{\link{summarizeBAMs}}; \code{\link{wigs2SE}}
}

