% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{ids2diffExpr}
\alias{ids2diffExpr}
\title{Run full differential expression pipeline}
\usage{
ids2diffExpr(ID, mnt_point = NULL)
}
\arguments{
\item{ID}{the vector of ContactExpIDs}

\item{mnt_point}{give the mount point for filesystem, use URLs if NULL, defaults to NULL}
}
\value{
stats string
}
\description{
Run full differential expression pipeline
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{bam2del}};
  \code{\link{doDiffExpr}};
  \code{\link{getDEreportForDESeqDataSet}};
  \code{\link{getDEreport}}; \code{\link{getDEseq2res}};
  \code{\link{getStage}}; \code{\link{getStrain}};
  \code{\link{getSummarizedEperiment}};
  \code{\link{summarizeBAMs}}
}

