% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jadb_cluster_ops.R
\name{jacl_send_to_cluster}
\alias{jacl_send_to_cluster}
\title{jacl_send_to_cluster}
\usage{
jacl_send_to_cluster(ID, genome = "ce11", ops = "", out_sufix = "chip",
  pipeline = "jadb_ChIPseq")
}
\arguments{
\item{ID}{}
}
\description{
jacl_send_to_cluster
}
\examples{
# sapply(JADB_mass_parallel(50), jacl_send_to_cluster)

# ids <- strsplit('AA652 AA653 AA654 AA655 AA656 AA657 AA658 AA659 AA660 AA661 AA662 AA663', ' ')[[1]]
# sapply(ids, jacl_send_to_cluster)

# ids <- jacl_mass_parallel(50)
# sapply(ids, jacl_send_to_cluster)

# jacl_send_to_cluster('AA007', ops=", steps=c(\\"log2_norm\\", \\"map0_norm\\", \\"log2_map0_norm\\"), purge=FALSE")

}
\seealso{
Other cluster: \code{\link{jacl_mass_parallel}},
  \code{\link{jacl_send_to_cluster_ce10}}
}
\author{
Przemyslaw Stempor
}
