% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_pipelenies.R
\name{jadb_ChIPseq}
\alias{jadb_ChIPseq}
\title{jadb_ChIPseq}
\usage{
jadb_ChIPseq(ids, steps = c("aln", "tracks", "norm", "fastqc", "macs"))
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
jadb_ChIPseq
}
\examples{
#
}
\author{
Przemyslaw Stempor
}
