% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_pipelenies.R
\name{jadb_ChIPseq}
\alias{jadb_ChIPseq}
\title{jadb_ChIPseq}
\usage{
jadb_ChIPseq(ids, steps = c("aln", "tracks", "norm", "log2_norm", "map0_norm",
  "log2_map0_norm", "map10_uniq", "fastqc", "fastqscreen", "macs", "meme"),
  genome = "ce11", purge = TRUE, backup = FALSE)
}
\arguments{
\item{IDs}{Vector of JADB ContactExpIDs}
}
\value{
List
}
\description{
jadb_ChIPseq
}
\examples{
#
}
\seealso{
Other pipelines: \code{\link{jadb_RNAseq}}
}
\author{
Przemyslaw Stempor
}
