% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikeInNorm.R
\name{normTrack}
\alias{normTrack}
\title{Normalize track by spike in ratio value ratios produced by}
\usage{
normTrack(nc)
}
\arguments{
\item{the}{named numeric representing normakization coefficient named with 
experiment ID in "XX001" format}
}
\value{
normalized track in BW format (connection)
}
\description{
The function produces normalized track using \code{\link{getRatio}} value.
}
\seealso{
Other spikeIn: \code{\link{getRatio}}
}
\author{
Przemyslaw Stempor
}
