% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNAseq.R
\name{summarizeBAMs}
\alias{summarizeBAMs}
\title{Get counts from bam files using embended gene meodel}
\usage{
summarizeBAMs(fls)
}
\arguments{
\item{fls}{file names vector}
}
\value{
eset
}
\description{
File URLs can be obtained like this:
fls <- sapply(c('rAM001', 'rAM002', 'rAM003', 'rAM004'), getFilePath, processing='aligned', format='bam')
}
\author{
Przemyslaw Stempor
}
\seealso{
Other RNAseq: \code{\link{bam2del}};
  \code{\link{doDiffExpr}};
  \code{\link{getDEreportForDESeqDataSet}};
  \code{\link{getDEreport}}; \code{\link{getDEseq2res}};
  \code{\link{getStrain}}; \code{\link{ids2diffExpr}}
}

